Checkpoint: everything up to add_deps_to_Wheel
The script now unpacks a wheel, searches within it for binary
artifacts, and identifies dependencies for each of those artifacts.
Then it pretends to add the dependencies to the wheel and repacks
it into a user-specified destination directory.

Next up: write add_dependencies_to_wheel().

After that the only big piece of functionality I know of is to
add HKLM\SYSTEM\CurrentControlSet\Control\Session Manager\KnownDLLs
to the search path.