Modify BuildWheel to take extra search paths
Since winlocate searches the user's PATH for files to satisfy
dependencies, we need a way to add directories to the PATH.
We can already do anything rooted at CMAKE_BINARY_DIR since that's
one of the arguments to the script but we also need a way to pass
in other things like the Boost search directory.

Solution: modify BuildWheel.cmake to take any arguments after the
required ones and interpret them as search paths.  Then, have
PythonWrapping/CMakeLists.txt pass in the Boost search directories
as arguments to the wheel target.