Fix argument name, docstring
I renamed the 'create_dir' argument to 'create_directory': according to the principle of least surprise, we want the user to not have to remember which things are abbreviated and which aren't wherever it makes sense.

Also moved the documentation for that argument out of Arguments and into Keyword Arguments.