Mention MSVC runtime, provide download link
Windows users on especially clean systems might need to install
the Microsoft Visual C++ runtime before Tracktable can work.  This
is now mentioned in the installation instructions along with a link
to the correct version.

This commit will be accompanied by a change in the FAQ on the web
site.