Trajectory constructor and test case for slicing
In Trajectory.h, created a new constructor that takes in two iterators and a trajectory as arguments.
Slices the trajectory based on the iterators and tranfers the properties from the original trajectory to the new sliced trajectory.

test_trajectory_slicing.cpp is a test case for the created constructor and tests the slicing works correctly and the newly sliced trajectory has the properties
of the original trajectory. (same values)

CMakeLists.txt calls the test case.

This will now allow the user to slice a trajectory by indices and be returned a trajectory object not a list of points. This is usefule for maintained perviously
established properties of the trajectory.