Fixed test_case for Pyhton testing of Trajectory Slicing
Modified test_trajectory_slicing - fixed a few syntax errors. The test case creates 3 random trajectories of different types (Cartesian2D, Cartesian3D
Terrestrial) and slices the trajectory into a smaller one. Tests that the smaller trajectory has the same properties as the original trajectory

Can now easily slice a trajectory in Python and C++ and be returned a trajectory (not a list of points)