Example of defining a docstring in one file and using it in another
This is just a placeholder to illustrate the concept I had in mind
when Mike and I discussed adding docstrings for domain classes in
Python.  Docstrings will need to be written in Python format (refer
to other classes in Tracktable's Python code or
https://www.python.org/dev/peps/pep-0257/, but be aware that the standards
have evolved since PEP0257 was written).

We need docstrings for all Python classes, methods, and functions.