Fix indentation error with docstring in render_trajectories
There was probably a tab loose somewhere.  Sphinx's error message was
not helpful.

I also fixed a whole bunch of cases where the existing code was using
'map' as the name of a variable.  While map is not a reserved word in
Python, it is in C++, and it seemed better to name it 'map_canvas'.
Still need to re-run Python tests to make sure I didn't break anything.