Replace obsolete code for getting platform, ABI tags
Our old schemes for retrieving the platform tag (cp37 for CPython
version 3.7) and ABI tag (cp37m for 'compiled with pymalloc') no longer
work due to changes in the modules "packaging" and "wheel".  This
latest approach tries to retrieve the ABI from the SOABI tag in
sysconfig and convert that.

If this breaks, we may have to punt and start hard-coding this to
'cp' for the C-Python interpreter.