Progress: package builds but test env fails
By switching to 'python setup.py install', I've got the library
building and installing in the right places.  I also added runtime
dependencies in meta.yaml.

The current problem is that conda fails when trying to create the
test environment due to conflicts.  Tracking that down should be
a matter of finding out where it's getting the package specifications
for the test environment.