Expose insert_point(), insert_points(), len() for R-tree
We have a feature request to be able to add points to an R-tree one
at a time.  I've added a binding to the C++ insert_point() method to
support this.  While I was at it I also added insert_points() for
completeness and rewrote the constructor to call that instead of doing
work itself.