Pull in TT-data changes that add DBF files to LFS
Wow, that's a lot of acronyms.

In order to store tracktable-data on Github, we have to move
all of our big files into Git LFS.  That's *.SHP and *.DBF
(the shapefile data) at present.

This commit on tracktable-data implements those changes.  I've
already pushed it successfully (and forcibly) both locally and
on Github, so we should be ready to go.

I might have to something similar with tracktable-docs due to the
big Jupyter notebooks.  We'll see.