Keys in gradient dict have to be strings
Found the bug!

When Folium is rendering things out to Javascript, it tries to convert
snake_case_variable_names to camelCaseVariableNames.  This gets applied
(deliberately or not) to the array containing the color gradient, whose
keys are floats.

The workaround (for now) is to make the keys strings so that they
get passed on to the javascript verbatim.