Update to mambaforge-latest, comment out git-lfs
On the trail of the no-notebooks bug.  I've updated .readthedocs.yml
to use mambaforge-latest in build.tools.python and commented out the
manual Git LFS install in the hope that RTD already has that in the
environment.

Next up: a bit of debug output in conf.py to figure out where we are
and see if the notebooks are there.