Resolve "FEATURE: Check that CI environment is up to date"
CI pipelines will now fail quickly if the environment file for the CI environment has been modified more recently than the environment itself.  This is both a security measure and a reality check -- if you've changed the dependencies in the environment file, CI would likely fail somewhere down the road.