Comment out up-to-date checks for CI environment
This check is too buggy.  It's causing spurious errors, saying that
the environment doesn't exist on Windows when I'm certain that it
does, and generating extra work due to merge commits.

Maybe in a future timeline I can resurrect it, make it into an
advisory instead of a failure, or even fix it.