Combine configure+build+test for CI
Summary: CMake build trees cannot be relocated, so we have to do configure, build, test on the same runner.  In an environment with multiple runners, that means we have to combine those three stages.

In truth, separating those three stages didn't give us any substantial benefit.  We could start to get an inkling of where things went wrong but we still had to inspect the logs to find details.